Move SassTransformer into its own src/sass.rs module
Keeps dist.rs focused on the Dist struct and Transformer trait.
sass.rs is compiled only when the sass feature is enabled and
re-exported via lib.rs with the usual cfg_attr(docsrs) gate.
dist.rs imports SassTransformer with #[cfg(feature = "sass")] for
the Default impl; the public API is unchanged.
